requirements.txt
spotipy
python-dotenv
unidecode
matplotlib